pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk-cost-tracker"
version = "0.1.0"
description = "Self-hosted, multi-provider LLM cost tracking for Google ADK, OpenAI, and AWS Bedrock"
readme = "README.md"
authors = [{ name = "STHITAPRAJNAS", email = "sthitaprajnas@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["llm", "cost", "tracking", "google-adk", "openai", "bedrock", "gemini", "observability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

# Zero hard dependencies — everything is optional to keep it lightweight
dependencies = []

[project.optional-dependencies]
postgres = ["asyncpg>=0.29"]
openai   = ["openai>=1.0"]
bedrock  = ["boto3>=1.34"]
google   = ["google-generativeai>=0.8"]
adk      = ["google-adk>=1.0"]
yaml     = ["PyYAML>=6.0"]
all      = [
    "asyncpg>=0.29",
    "openai>=1.0",
    "boto3>=1.34",
    "google-generativeai>=0.8",
    "google-adk>=1.0",
    "PyYAML>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
adk-cost-report = "adk_cost_tracker.report:main"

[project.urls]
Homepage   = "https://github.com/STHITAPRAJNAS/adk_cost_tracker"
Repository = "https://github.com/STHITAPRAJNAS/adk_cost_tracker"
Issues     = "https://github.com/STHITAPRAJNAS/adk_cost_tracker/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_cost_tracker"]

# ── pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
addopts      = "-v --tb=short"

# ── ruff (linting) ────────────────────────────────────────────────────────
[tool.ruff]
line-length    = 100
target-version = "py311"
select         = ["E", "F", "I", "UP", "B"]

# ── mypy ──────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict         = false
ignore_missing_imports = true
